// types.ts
export interface Transaction {
  id: string;
  userId: string;
  userName: string;
  type: 'deposit' | 'withdrawal' | 'payout' | 'referral_bonus' | 'task_reward';
  amount: number;
  status: 'pending' | 'completed' | 'rejected';
  paymentMethod?: string;
  accountDetails?: string;
  proofUrl?: string; // Mock uploaded proof image
  gatewayReference?: string;
  gatewayChannel?: string;
  createdAt: string;
  description: string;
  isMarketing?: boolean;
}